Initial commit of credential tester
Help troubleshoot a CLI issue with '+' and '/' chars.
The idea here is to write the same script in different languages
and see what happens.  If it fails in both languages it's likely
an issue with the credentials themselves.  If it's an issue
with a single SDK, then we should see the other SDKs pass while the
affected SDK fails.